fix: signup 파일 구조 변경 (#281)
