--- frontend/src/components/Common/Progressbar/index.jsx
@@ -1,35 +1,0 @@
-import React from "react";
-
-const ProgressBar = (props) => {
-  const { bgcolor, completed } = props;
-
-  const containerStyles = {
-    height: 5,
-    width: "100%",
-    backgroundColor: "#e0e0de",
-  };
-
-  const fillerStyles = {
-    height: "100%",
-    width: `${completed}%`,
-    backgroundColor: bgcolor,
-    borderRadius: "inherit",
-    textAlign: "right",
-  };
-
-  const labelStyles = {
-    padding: 5,
-    color: "white",
-    fontWeight: "bold",
-  };
-
-  return (
-    <div style={containerStyles}>
-      <div style={fillerStyles}>
-        <span style={labelStyles}></span>
-      </div>
-    </div>
-  );
-};
-
-export default ProgressBar;
