--- frontend/src/components/SignupInfo/index.jsx
@@ -1,1 +1,0 @@
-export { default as SignupInfoForm } from "./SignupInfoForm";
