+++ frontend/src/pages/SignUP/index.jsx
@@ -3,7 +3,7 @@ import MainComponent from "../../components/Main";
 import SignUpComponent from "../../components/SignUp";
 import { NavBar } from "../../components/Common";
 
-export default function SignupInfo() {
+export default function SignUp() {
   return (
     <>
       <NavBar />
